Rename `assistant_settings` to `agent_settings`
