+++ crates/agent_settings/LICENSE-GPL
