+++ crates/agent_settings/src/agent_profile.rs
