+++ crates/assistant_context_editor/Cargo.toml
@@ -12,8 +12,8 @@ workspace = true
 path = "src/assistant_context_editor.rs"
 
 [dependencies]
+agent_settings.workspace = true
 anyhow.workspace = true
-assistant_settings.workspace = true
 assistant_slash_command.workspace = true
 assistant_slash_commands.workspace = true
 chrono.workspace = true
