+++ crates/collab/Cargo.toml
@@ -76,8 +76,8 @@ workspace-hack.workspace = true
 zed_llm_client.workspace = true
 
 [dev-dependencies]
+agent_settings.workspace = true
 assistant_context_editor.workspace = true
-assistant_settings.workspace = true
 assistant_slash_command.workspace = true
 assistant_tool.workspace = true
 async-trait.workspace = true
