+++ crates/eval/src/examples/code_block_citations.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use assistant_settings::AgentProfileId;
+use agent_settings::AgentProfileId;
 use async_trait::async_trait;
 use markdown::PathWithRange;
