+++ crates/eval/src/examples/file_search.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use assistant_settings::AgentProfileId;
+use agent_settings::AgentProfileId;
 use assistant_tools::FindPathToolInput;
 use async_trait::async_trait;
 use regex::Regex;
