+++ crates/zed/Cargo.toml
@@ -17,11 +17,11 @@ path = "src/main.rs"
 [dependencies]
 activity_indicator.workspace = true
 agent.workspace = true
+agent_settings.workspace = true
 anyhow.workspace = true
 askpass.workspace = true
 assets.workspace = true
 assistant_context_editor.workspace = true
-assistant_settings.workspace = true
 assistant_tool.workspace = true
 assistant_tools.workspace = true
 async-watch.workspace = true
@@ -42,10 +42,10 @@ command_palette.workspace = true
 component.workspace = true
 copilot.workspace = true
 dap_adapters.workspace = true
-debugger_ui.workspace = true
-debugger_tools.workspace = true
 db.workspace = true
 debug_adapter_extension.workspace = true
+debugger_tools.workspace = true
+debugger_ui.workspace = true
 diagnostics.workspace = true
 editor.workspace = true
 env_logger.workspace = true
@@ -80,8 +80,8 @@ language_tools.workspace = true
 languages = { workspace = true, features = ["load-grammars"] }
 libc.workspace = true
 log.workspace = true
-markdown_preview.workspace = true
 markdown.workspace = true
+markdown_preview.workspace = true
 menu.workspace = true
 migrator.workspace = true
 mimalloc = { version = "0.1", optional = true }
@@ -141,12 +141,12 @@ vim_mode_setting.workspace = true
 web_search.workspace = true
 web_search_providers.workspace = true
 welcome.workspace = true
+workspace-hack.workspace = true
 workspace.workspace = true
 zed_actions.workspace = true
 zeta.workspace = true
 zlog.workspace = true
 zlog_settings.workspace = true
-workspace-hack.workspace = true
 
 [target.'cfg(target_os = "windows")'.dependencies]
 windows.workspace = true
